remove extra logs, more comments
