+++ azure_data_cosmos_engine/src/query/node.rs
@@ -67,7 +67,6 @@ impl<'a> PipelineSlice<'a> {
             None => {
                 tracing::debug!("retrieving item from producer");
                 let result = self.producer.produce_item()?;
-                tracing::debug!(?result, "completed item production");
                 Ok(result)
             }
         }
